+++ src/lib/abstraction/form-builder.ts
@@ -1,17 +1,13 @@
 import {AbstractGroupType, IGroupTypeOptions} from '../form-type/abstract-group-type';
 import {FormModel} from '../form-type/abstract-type';
-import {EventEmitter, QueryList} from '@angular/core';
-import {FormSlotDirective} from '../form-slot/form-slot.directive';
-import {FormGroup} from './form-group';
-import {FormArray} from './form-array';
-import {ValidationErrors} from '@angular/forms';
+import {EventEmitter} from '@angular/core';
+import {FormArray, FormGroup, ValidationErrors} from '@angular/forms';
 
 export interface IFormBuilder<T extends { [key: string]: any } = {}> {
   fieldType: AbstractGroupType<IGroupTypeOptions<T>, T>;
   mwFormModel: FormModel<T>;
   mwFormData: T;
   mwFormSubmit: EventEmitter<T>;
-  mwSlots: QueryList<FormSlotDirective>;
   group: FormGroup;
   readonly valid: boolean;
   readonly invalid: boolean;