+++ src/lib/form-builder.module.ts
@@ -3,7 +3,6 @@ import {FormBuilderComponent} from './form-builder/form-builder.component';
 import {ReactiveFormsModule} from '@angular/forms';
 import {FormFieldDirective} from './form-field/form-field.directive';
 import {CommonModule} from '@angular/common';
-import {FormSlotDirective} from './form-slot/form-slot.directive';
 import {FormGroupComponent} from './form-group/form-group.component';
 import {DummyFieldComponent} from './dummy-field/dummy-field.component';
 
@@ -12,7 +11,6 @@ import {DummyFieldComponent} from './dummy-field/dummy-field.component';
     FormBuilderComponent,
     FormGroupComponent,
     FormFieldDirective,
-    FormSlotDirective,
     DummyFieldComponent,
   ],
   entryComponents: [
@@ -26,7 +24,6 @@ import {DummyFieldComponent} from './dummy-field/dummy-field.component';
     FormBuilderComponent,
     FormGroupComponent,
     FormFieldDirective,
-    FormSlotDirective,
   ],
 })
 export class FormBuilderModule {