+++ src/lib/form-field/form-field.directive.ts
@@ -1,5 +1,4 @@
-import {ComponentFactoryResolver, Directive, ElementRef, ViewContainerRef} from '@angular/core';
-import {OnChanges, SimpleChanges} from '@angular/core';
+import {ComponentFactoryResolver, Directive, ElementRef, OnChanges, SimpleChanges, ViewContainerRef} from '@angular/core';
 import {AbstractFormFieldComponent} from '../abstract-form-field/abstract-form-field.component';
 import {AbstractGroupType} from '../form-type/abstract-group-type';
 import {FormGroup} from '@angular/forms';
@@ -37,7 +36,6 @@ export class FormFieldDirective extends AbstractFormFieldComponent<any> implemen
       mwFormGroup: isGroup ? this.mwElement as FormGroup : this.mwFormGroup,
       mwElement: this.mwElement,
       mwFieldType: this.mwFieldType,
-      mwSlots: this.mwSlots,
       mwPath: this.mwPath,
       mwIndex: this.mwIndex,
     });